Extract applying keywords to separate game step
