+++ server/game/gamesteps/challenge/challengeflow.js
@@ -5,7 +5,7 @@ const SimpleStep = require('../simplestep.js');
 const ChooseStealthTargets = require('./choosestealthtargets.js');
 const ApplyClaim = require('./applyclaim.js');
 const ActionWindow = require('../actionwindow.js');
-const GameKeywords = require('../../gamekeywords.js');
+const KeywordWindow = require('../keywordwindow.js');
 
 class ChallengeFlow extends BaseStep {
     constructor(game, challenge) {
@@ -26,7 +26,7 @@ class ChallengeFlow extends BaseStep {
             new SimpleStep(this.game, () => this.determineWinner()),
             new SimpleStep(this.game, () => this.unopposedPower()),
             new SimpleStep(this.game, () => this.beforeClaim()),
-            new SimpleStep(this.game, () => this.applyKeywords()),
+            () => new KeywordWindow(this.game, this.challenge),
             new SimpleStep(this.game, () => this.completeChallenge())
         ]);
     }
@@ -249,35 +249,6 @@ class ChallengeFlow extends BaseStep {
         this.game.addMessage('{0} continues without applying claim', player, this);
 
         return true;
-    }
-
-    applyKeywords() {
-        const challengeKeywords = ['insight', 'intimidate', 'pillage', 'renown'];
-        let winnerCards = this.challenge.getWinnerCards();
-        let appliedIntimidate = false;
-
-        _.each(winnerCards, card => {
-            let context = { game: this.game, challenge: this.challenge, source: card };
-
-            _.each(challengeKeywords, keyword => {
-                // It is necessary to check whether intimidate has been applied
-                // here instead of in the ability class because the individual
-                // keywords are resolved asynchronously but are queued up
-                // synchronously here. So two intimidates could be queued when
-                // only one is allowed.
-                if(keyword === 'intimidate' && appliedIntimidate) {
-                    return;
-                }
-
-                let ability = GameKeywords[keyword];
-                if(card.hasKeyword(keyword) && ability.meetsRequirements(context)) {
-                    appliedIntimidate = appliedIntimidate || (keyword === 'intimidate');
-                    this.game.resolveAbility(ability, context);
-                }
-            });
-
-            this.game.checkWinCondition(this.challenge.winner);
-        });
     }
 
     completeChallenge() {
